+++ kpi/tests/test_asset_content.py
@@ -6,7 +6,8 @@ from kpi.models import Asset
 from kpi.utils.sluggify import sluggify_label
 from pprint import pprint
 from collections import OrderedDict
-import pytest
+
+# import pytest
 import json
 import string
 import inspect
@@ -49,6 +50,7 @@ def rank_asset_content():
         ],
         u'settings': {},
     }
+
 
 def rank_asset_named_content():
     return {
@@ -449,3 +451,97 @@ def test_named_score_question_compiles():
         [u'end_group', False,
             u'$kuid']
     ]
+
+
+def kobomatrix_content():
+    return {
+        u'survey': [
+            {u'type': 'begin_kobomatrix',
+                'name': 'm1',
+                'label': 'Items',
+                'kobo--matrix_list': 'car_bike_tv',
+                },
+            {u'type': 'select_one', 'select_from_list_name': 'yn',
+              'label': 'Possess?', 'name': 'possess', 'required': True},
+            {u'type': 'select_one', 'select_from_list_name': 'yn',
+              'label': 'Necessary?', 'name': 'necess', 'required': True},
+            {u'type': 'integer',
+              'label': 'Number', 'name': 'number', 'required': True},
+            {u'type': 'end_kobomatrix'},
+        ],
+        u'choices': [
+            {u'list_name': 'car_bike_tv', u'label': 'Car', 'name': 'car'},
+            {u'list_name': 'car_bike_tv', u'label': 'Bike', 'name': 'bike'},
+            {u'list_name': 'car_bike_tv', u'label': 'TV', 'name': 'tv'},
+        ],
+        u'settings': {},
+    }
+
+
+def test_kobomatrix_content():
+    content = _compile_asset_content(kobomatrix_content())
+    pattern = ['w7', 'w1', 'w2', 'w2', 'w2', None]
+    _survey = content.get('survey')
+    _names = [r.get('name') for r in _survey]
+    _labls = [r.get('label', [None])[0] for r in _survey]
+    _none_labels = [label is None for label in _labls]
+    _reqds = [r.get('required', None) for r in _survey]
+
+    # appearance fields match
+    assert [r.get('appearance') for r in _survey] == pattern * 4
+
+    # names match
+    assert _names == ['m1_header',
+                      'm1_header_note',
+                      'm1_header_possess',
+                      'm1_header_necess',
+                      'm1_header_number',
+                      None,
+                      'm1_car',
+                      'm1_car_note',
+                      'm1_car_possess',
+                      'm1_car_necess',
+                      'm1_car_number',
+                      None,
+                      'm1_bike',
+                      'm1_bike_note',
+                      'm1_bike_possess',
+                      'm1_bike_necess',
+                      'm1_bike_number',
+                      None,
+                      'm1_tv',
+                      'm1_tv_note',
+                      'm1_tv_possess',
+                      'm1_tv_necess',
+                      'm1_tv_number',
+                      None,
+                      ]
+
+    assert _none_labels == [True, False, False, False, False, True] * 4
+    assert _labls[1:5] == ['**Items**',
+                           '**Possess?**',
+                           '**Necessary?**',
+                           '**Number**',
+                           ]
+
+    def _span(item):
+        return '<span style="display:none">{}</span>'.format(item)
+
+    assert _labls[7:11] == ['##### Car',
+                            _span('car-Possess?'),
+                            _span('car-Necessary?'),
+                            _span('car-Number'),
+                            ]
+    assert _labls[13:17] == ['##### Bike',
+                             _span('bike-Possess?'),
+                             _span('bike-Necessary?'),
+                             _span('bike-Number'),
+                             ]
+    assert _labls[19:23] == ['##### TV',
+                             _span('tv-Possess?'),
+                             _span('tv-Necessary?'),
+                             _span('tv-Number'),
+                             ]
+    assert _reqds == [None, False, False, False, False, None] + (
+                        [None, False, True, True, True, None] * 3
+                    )
